pyproject.toml
[tool.poetry]
name = "aws_lambdas_deploy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9.6,<3.11"
scikit-learn = "^1.0.2"
numpy = "^1.21.5"
pandas = "^1.3.5"
joblib = "^1.1.0"
requests = "^2.26.0"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
boto3 = "^1.20.26"
matplotlib = "^3.5.1"
pyjanitor = "^0.22.0"
jupyter = "^1.0.0"
ipykernel = "^6.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
